#46872 - Move stories to components directories
