+++ .storybook/config.js
@@ -2,7 +2,7 @@
 import { configure } from '@storybook/vue';
 
 function loadStories() {
-  const req = require.context('../stories', true, /.stories.js$/);
+  const req = require.context('../components', true, /.stories.js$/);
   req.keys().forEach(filename => req(filename));
 }
