+++ components/01-globals/app/App.stories.js
@@ -1,6 +1,6 @@
 import { storiesOf } from '@storybook/vue'
 
-import App from '../components/01-globals/app/App.vue'
+import App from './App.vue'
 
 storiesOf('Global/App', module)
   .add('Default', () => ({
