+++ components/01-globals/container/Container.stories.js
@@ -1,7 +1,7 @@
 import { storiesOf } from '@storybook/vue'
 
-import App from '../components/01-globals/app/App.vue'
-import Container from '../components/01-globals/container/Container.vue'
+import App from '../app/App.vue'
+import Container from './Container.vue'
 
 storiesOf('Global/Container', module)
   .add('Default', () => ({
