+++ components/03-modules/header/03-modules-header.stories.js
@@ -1,7 +1,7 @@
 import { storiesOf } from '@storybook/vue'
 
-import App from '../components/01-globals/app/App.vue'
-import AlpacaHeader from '../components/03-modules/header/Header.vue'
+import App from '../../01-globals/app/App.vue'
+import AlpacaHeader from './Header.vue'
 
 storiesOf('Modules/Header', module)
   .add('Default', () => ({
